// Frontend/ClientApp/src/app/models/SeriesEmision/SeriesEmision.ts
export class SeriesEmisionC{
  Name: string = "Dinámico";
  AuthorizationNumber: string = "18760000001";
  Prefix: string = "";
  ValidFrom: Date = new Date("2019-01-19T15:00:00");
  ValidTo: Date = new Date("2030-01-19T05:00:00");
  StartValue: number = 1;
  EndValue: number = 5000000;
  EfectiveValue: any = null;
  DocumentType: string = "";
  TechnicalKey: string = "fc8eac422eba16e22ffd8c6f94b3f40a6e38162c";
  AutoIncrement: boolean = false;
}